+++ build.gradle
@@ -25,7 +25,7 @@ ext {
 apply plugin: 'application'
 apply from: 'gradle/formatting.gradle'
 
-mainClassName = 'opensearchSDK.ExtensionsRunner'
+mainClassName = 'org.opensearch.sdk.ExtensionsRunner'
 
 group 'org.example'
 version '1.0-SNAPSHOT'
