+++ src/main/java/org/opensearch/sdk/ExtensionsRunner.java
@@ -9,7 +9,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK;
+package org.opensearch.sdk;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -39,8 +39,8 @@ import org.opensearch.transport.ConnectionManager;
 import org.opensearch.transport.TransportService;
 import org.opensearch.transport.TransportSettings;
 
-import opensearchSDK.transport.netty4.Netty4Transport;
-import opensearchSDK.transport.netty4.SharedGroupFactory;
+import org.opensearch.sdk.netty4.Netty4Transport;
+import org.opensearch.sdk.netty4.SharedGroupFactory;
 
 import org.opensearch.transport.TransportInterceptor;
