+++ src/main/java/org/opensearch/sdk/netty4/CopyBytesSocketChannel.java
@@ -44,7 +44,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport.netty4;
+package org.opensearch.sdk.netty4;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
