+++ src/main/java/org/opensearch/sdk/netty4/Netty4MessageChannelHandler.java
@@ -30,7 +30,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport.netty4;
+package org.opensearch.sdk.netty4;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.*;
