+++ src/main/java/org/opensearch/sdk/netty4/Netty4NioSocketChannel.java
@@ -15,7 +15,7 @@
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -30,27 +30,29 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport.netty4;
+package org.opensearch.sdk.netty4;
 
-import io.netty.buffer.ByteBuf;
-import io.netty.channel.ChannelHandler;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.handler.codec.MessageToMessageDecoder;
+import io.netty.channel.Channel;
+import io.netty.channel.socket.nio.NioSocketChannel;
 
-import java.util.List;
+import java.nio.channels.SocketChannel;
 
-@ChannelHandler.Sharable
-public class NettyByteBufSizer extends MessageToMessageDecoder<ByteBuf> {
+/**
+ * Helper class to expose {@link #javaChannel()} method
+ */
+public class Netty4NioSocketChannel extends NioSocketChannel {
+
+    public Netty4NioSocketChannel() {
+        super();
+    }
+
+    public Netty4NioSocketChannel(Channel parent, SocketChannel socket) {
+        super(parent, socket);
+    }
 
     @Override
-    protected void decode(ChannelHandlerContext ctx, ByteBuf buf, List<Object> out) {
-        int readableBytes = buf.readableBytes();
-        if (buf.capacity() >= 1024) {
-            ByteBuf resized = buf.discardReadBytes().capacity(readableBytes);
-            assert resized.readableBytes() == readableBytes;
-            out.add(resized.retain());
-        } else {
-            out.add(buf.retain());
-        }
+    public SocketChannel javaChannel() {
+        return super.javaChannel();
     }
+
 }
