+++ src/main/java/org/opensearch/sdk/netty4/Netty4TcpServerChannel.java
@@ -30,7 +30,7 @@
  * GitHub history for details.
  */
 
-package opensearchSDK.transport.netty4;
+package org.opensearch.sdk.netty4;
 
 import io.netty.channel.Channel;
 import org.opensearch.action.ActionListener;
